backend/pyproject.toml
[project]
name = "req2veri-backend"
version = "0.1.0"
description = "FastAPI backend for req2veri"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.17",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "httpx>=0.27.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
